Practice what we preach. Hoorah for portability.
Internal counters switched to size_t (tableio_size_t == uint64_t when written 
out to disk, so that fields are the same size no matter where they're written).


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/tableio@61800 16731396-06f5-0310-8873-f7f720988828

+++ private/pybindings/I3Converter.cxx
@@ -24,7 +24,7 @@ void register_I3Converter() {
 		(I3TableRowDescriptionConstPtr (PythonConverter::*)()) &PythonConverter::GetDescription)
 	.def("CreateDescription",(I3TableRowDescriptionPtr (PythonConverter::*)(I3FrameObjectConstPtr))
                     &PythonConverter::CreateDescription)
-	.def("Convert",(unsigned int (PythonConverter::*)(I3FrameObjectConstPtr, 
+	.def("Convert",(size_t (PythonConverter::*)(I3FrameObjectConstPtr, 
 	                                                  I3TableRowPtr, 
 	                                                  I3FramePtr))
 	               &PythonConverter::Convert)
