+++ private/tableio/PythonConverter.h
@@ -9,7 +9,7 @@ struct PythonConverter : I3Converter, boost::python::wrapper<I3Converter>
 
     // void Configure( const StringPairVector& params);
 
-    unsigned int GetNumberOfRows(I3FrameObjectConstPtr object);
+    size_t GetNumberOfRows(I3FrameObjectConstPtr object);
 
     I3TableRowDescriptionConstPtr GetDescription(I3FrameObjectConstPtr object);
     I3TableRowDescriptionConstPtr GetDescription(const I3FrameObject& object);
@@ -19,15 +19,15 @@ struct PythonConverter : I3Converter, boost::python::wrapper<I3Converter>
     I3TableRowDescriptionPtr CreateDescription(I3FrameObjectConstPtr object);
     I3TableRowDescriptionPtr CreateDescription(const I3FrameObject& object);
     
-    unsigned int Convert(I3FrameObjectConstPtr object, 
+    size_t Convert(I3FrameObjectConstPtr object, 
                                  I3TableRowPtr rows, 
                                  I3FramePtr frame=I3FramePtr());
     
-    unsigned int Convert(const I3FrameObject& object, 
+    size_t Convert(const I3FrameObject& object, 
                                  I3TableRowPtr rows, 
                                  I3FramePtr frame=I3FramePtr());
 
-    unsigned int FillRows(const I3FrameObjectConstPtr object, I3TableRowPtr rows);
+    size_t FillRows(const I3FrameObjectConstPtr object, I3TableRowPtr rows);
 
     bool CanConvert(I3FrameObjectConstPtr object);
     bool CanConvert(I3FrameObjectPtr object);
