+++ public/tableio/internals/I3Table.h
@@ -31,23 +31,23 @@ class I3Table {
         bool IsConnectedToWriter();
         void SetConnectedToWriter(bool connected);
 
-        I3TableRowPtr CreateRow(unsigned int nrows);
+        I3TableRowPtr CreateRow(size_t nrows);
         void AddRow(I3EventHeaderConstPtr header, I3TableRowConstPtr row);
         void Align();
         
         I3TableRowConstPtr GetRowForEvent(unsigned int RunID, unsigned int EventID);
-        I3TableRowConstPtr GetRowForEvent(unsigned int index);
+        I3TableRowConstPtr GetRowForEvent(size_t index);
 
         std::string GetName() const; 
-        unsigned int GetNumberOfEvents() const; 
-        unsigned int GetNumberOfRows() const; 
+        size_t GetNumberOfEvents() const; 
+        size_t GetNumberOfRows() const; 
         I3TableRowDescriptionConstPtr GetDescription();
 
     protected:
         // to be overridden by derivatives
         virtual void WriteRows(I3TableRowConstPtr row) = 0;
         virtual I3TableRowConstPtr ReadRows(size_t start, size_t nrows);
-        virtual std::pair<unsigned int,unsigned int> GetRangeForEvent(unsigned int index);
+        virtual std::pair<size_t,size_t> GetRangeForEvent(size_t index);
 
         I3TableService& service_;
         std::string name_;
@@ -55,9 +55,9 @@ class I3Table {
         I3TableRowDescriptionConstPtr description_;
         boost::shared_ptr<I3Table> indexTable_;
 
-        unsigned int nevents_; // number of events written
-        unsigned int nrows_;   // number of rows (>= events) written
-        unsigned int nrowsWithPadding_;   // number of rows (>= events) written, including padding
+        size_t nevents_; // number of events written
+        size_t nrows_;   // number of rows (>= events) written
+        size_t nrowsWithPadding_;   // number of rows (>= events) written, including padding
         bool connected_;       // connected to a writer
         bool tableCreated_;    // the table/tree has been created successfully
